Remove CPU/Memory Usage
These metrics come from metrics that are deployment-specific.
Openshift will provide:
 + container_memory_working_set_bytes
 + container_cpu_usage_seconds_total

It is more difficult to get the same data for a containerized
deployment that spans across multiple nodes.

The decision to remove those visualizations from this dashboard
comes with the idea that this dashboard should be trated as a
starting point from which the user will extend to its needs.